Make it clear to user if an order has been overridden/authorized by DH
Marks an order on the order list page.  The result is that if a User creates an order for a meal they aren't eligible for
the order will show not eligible until it is overridden/authorized and then it will show Order Authorized.
This increase in clarity should reduce questions and confusion