Force DH to look at Default Contact info page for their Dept if blank
Upon loading the Dept_order page if nothing is configured in the default contact info fields for the
department a user is working in system will redirect to the page to add this.
It will allow the DH to go back to the Dept Order page without adding any contact info.
It will also redirect the user back to this page every time they click on something that loads the Dept Order screen
until someone puts something in.
Hopefully this will result in people filling this out with useful info rather than putting trash.