Added webhook notification option for DH or Admin users
Added an option for webhook style notification for DH or admin users.
I have limited it to these because I don't know enough about webhooks to be confident this couldn't be abused
and because I have vague fears that allowing just any user to have webhook notifications may cause confusion with runners.
Also, why would anyone but a DH need this?  If they want to know just hang out in Slack right?