Add Food Manager feature
A food manager is a non-dh person who is assigned to handle orders for other people.
They have the same access as a DH but have to be manually added to a list of food managers by an existing DH or Admin.
The box to add them is on the Dept Orders screen after you have selected a specific department,
for lack of a better place.