Update SMS and Email errors to DM me
was going to public Slack channel, need it to show who it was trying to
reach and don't want that to be in a public place.

Also fixed bugs in error handling for AWS,
before if it got an error that would kill the whole loop
instead of just continuing the loop after handling the error.