Remove order specific contact details
Previously there were both default contact details and the option to add separate order specific contact details that only apply to a specific order and override the default ones for just that order.  This caused much confusion for people who only use the app one week out of the year, so I have removed this.