Allow the app to track attendees by badge number instead of public ID
As of the time of this commit Uber API is not allowing Attendee Search or Lookup with the attendee's Public ID.  The app is designed around tracking everything by their public ID, but luckily most of the system doesn't actually require that and will work with the badge number with only minor modifications.
This adds a switch in the login code (not automatic or in config at this time) to allow using badge number instead of public ID, and the necessary modifications elsewhere to allow the rest of the system to handle what it sees as an int.