Upgrade Dept Contact Info system
Department contact info now has a separate page, rather than being stuffed in the Dept Order screen.
The previous arrangement was intended to allow setting meal-specific contact info in addition to the default contact info (with the theory that different people are on shift at different times and may want different contacts) but this was more complicated than it needed to be and caused confusion.